requirements.txt
Jinja2
pydantic
beautifulsoup4
lxml
python-slugify